tag search

+++ types/schemasearch.go
@@ -16,9 +16,9 @@ const (
 )
 
 type SchemaSearchRequest struct {
-	UserID         *float64            `json:"user_id"`
-	SchemaID       *float64            `json:"schema_id"`
-	TagID          *float64            `json:"tag_id"`
+	UserID         *int64              `json:"user_id"`
+	SchemaID       *int64              `json:"schema_id"`
+	TagID          *int64              `json:"tag_id"`
 	SchemaName     *string             `json:"schema_name"`
 	UserName       *string             `json:"user_name"`
 	Keywords       *string             `json:"keywords"`
